Draft mechanism to refresh only if display needs to change
It’s primitive though and needs serious code cleanup.

Notably, need to make a deep copy of the nodes JsonObject that is
passed to the model. Deleting elements currently removes them from
the data used to construct the display, so e.g. deleting battery
means it’s no longer displayed.